+++ libs/data_structures/matrix/matrix.h
@@ -48,7 +48,7 @@ void matrix_swap_columns(matrix m, int j1, int j2);
 
 // Выполняет сортировку вставками строк матрицы m по неубыванию значения
 // функции criteria применяемой для строк
-void matrix_insertion_sort_rows_by_row_criteria(matrix m, int (*criteria)(int*, int));
+void matrix_insertion_sort_rows_by_row_criteria(matrix m, float (*criteria)(int*, int));
 
 // Выполняет сортировку выбором столбцов матрицы m по неубыванию значения
 // функции criteria применяемой для столбцов
